# client/public/agent-access.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blueprint Robot-Team Agent API",
    "version": "2026-05-21",
    "summary": "Headless catalog, site-world, dry-run commerce, hosted-session, explorer-render, and export contract for robot-team agents.",
    "description": "Blueprint exposes capture-backed site-world catalog, dry-run quote/order/entitlement proof, and hosted-session endpoints for robot-team agents. Public demo calls may run without privileged credentials only when a demo site world is enabled. Protected robot-team flows use the existing Firebase bearer path and require robot_team/admin outer auth plus session ownership or a matching provisioned entitlement."
  },
  "servers": [
    {
      "url": "https://tryblueprint.io",
      "description": "Production public origin. Protected flows require a scoped robot-team/admin bearer token."
    },
    {
      "url": "http://localhost:5000",
      "description": "Local development server. Public demo flow can run without auth when demo site worlds are enabled."
    }
  ],
...
  "paths": {
    "/api/agent-access": {
      "get": {
        "tags": [
          "Discovery"
        ],
        "operationId": "discoverAgentAccess",
        "summary": "Discover the robot-team agent access manifest.",
        "description": "Credential-free discovery manifest for headless robot-team agents. It names blueprint.siteWorld.search as the first-class search tool, keeps requestCandidate intake-only, lists dry-run commerce endpoints, documents truth labels, and explains the mock/public-demo hosted-session path without credentials.",
        "security": [
          {}
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAccessManifest"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agent-access/commerce/quote": {
      "get": {
        "tags": [
          "Agent commerce"
        ],
        "operationId": "quoteAgentCommerce",
        "summary": "Quote a site-world package or hosted-session rental for agent planning.",
        "security": [
          {}
        ],
        "parameters": [
          {
            "name": "siteWorldId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "product",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "site_world_package",
                "hosted_session_rental"
              ],
              "default": "hosted_session_rental"
            }
          },
          {
            "name": "sessionHours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCommerceQuoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-blueprint-dry-run-only": true,
        "x-blueprint-truth-boundary": "Quotes are planning artifacts. They do not create live Stripe sessions, charge cards, grant live package access, or prove rights clearance."
      }
    },
    "/api/agent-access/commerce/dry-run-checkout": {
      "post": {
        "tags": [
          "Agent commerce"
        ],
        "operationId": "createAgentDryRunCheckout",
        "summary": "Create a dry-run order and provisioned entitlement without live Stripe.",
        "security": [
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDryRunCheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dry-run order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDryRunOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-blueprint-dry-run-only": true,
        "x-blueprint-truth-boundary": "This route reuses buyerOrders and marketplaceEntitlements response shapes for local/test proof only. It never calls live Stripe or opens live package access."
      }
    },
    "/api/agent-access/commerce/orders/{orderId}": {
      "get": {
        "tags": [
          "Agent commerce"
        ],
        "operationId": "getAgentDryRunOrder",
        "summary": "Read a dry-run order and receipt.",
        "security": [
          {}
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDryRunOrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Requested site world or session was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-blueprint-dry-run-only": true
      }
    },
    "/api/agent-access/commerce/entitlements/{entitlementId}": {
      "get": {
        "tags": [
          "Agent commerce"
        ],
        "operationId": "getAgentDryRunEntitlement",
        "summary": "Read a dry-run marketplace entitlement.",
        "security": [
          {}
        ],
        "parameters": [
          {
            "name": "entitlementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentEntitlementResponse"
                }
              }
            }
          },
          "404": {
            "description": "Requested site world or session was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-blueprint-dry-run-only": true
      }
    },
    "/api/agent-access/commerce/entitlement-readiness": {
      "get": {
        "tags": [
          "Agent commerce"
        ],
        "operationId": "getAgentEntitlementReadiness",
        "summary": "Check whether a provisioned entitlement would unlock protected hosted-session launch.",
        "security": [
          {}
        ],
        "parameters": [
          {
            "name": "siteWorldId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entitlementId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "buyerUserId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "agent-dry-run-buyer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentEntitlementReadiness"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-blueprint-dry-run-only": true,
        "x-blueprint-truth-boundary": "Entitlement readiness proves commerce linkage only. Runtime launch, rights, provider execution, and package fulfillment still require their owning systems."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BlueprintBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token or future scoped robot-team machine token",
        "description": "Protected flows currently use Firebase Admin token verification. The bearer must resolve to an admin user or a user profile with buyerType robot_team."
      }
    },
...
    "schemas": {
      "TruthLabel": {
        "type": "string",
        "enum": [
          "capture_grounded",
          "provider_derived",
          "generated",
          "sample_demo",
          "public_demo_eligible",
          "request_gated",
          "protected_robot_team",
          "dry_run_order"
        ]
      },
      "AgentCommerceProduct": {
        "type": "string",
        "enum": [
          "site_world_package",
          "hosted_session_rental"
        ]
      },
...
      "AgentAccessManifest": {
        "type": "object",
        "required": [
          "name",
          "preferredTool",
          "compatibilityTool",
          "credentiallessWorkflow",
          "siteWorldSearch",
          "requestCandidate",
          "dryRunCommerce",
          "publicDemo",
          "protectedFlow",
          "truthLabels"
        ],
        "additionalProperties": true,
        "properties": {
          "name": {
            "type": "string"
          },
          "docs": {
            "type": "string"
          },
          "openapi": {
            "type": "string"
          },
          "staticOpenapi": {
            "type": "string"
          },
          "llms": {
            "type": "string"
          },
          "llmsFull": {
            "type": "string"
          },
          "preferredTool": {
            "type": "string",
            "enum": [
              "blueprint.siteWorld.search"
            ]
          },
          "compatibilityTool": {
            "type": "string",
            "enum": [
              "blueprint.catalog.search"
            ]
          },
          "publicDemoSiteWorldId": {
            "type": "string"
          },
          "credentiallessWorkflow": {
            "type": "object",
            "required": [
              "requiredCredentials",
              "summary",
              "smokeCommand",
              "cliCommands"
            ],
            "additionalProperties": true,
            "properties": {
              "requiredCredentials": {
                "type": "boolean"
              },
              "summary": {
                "type": "string"
              },
              "smokeCommand": {
                "type": "string"
              },
              "cliCommands": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "siteWorldSearch": {
            "type": "object",
            "additionalProperties": true
          },
          "requestCandidate": {
            "type": "object",
            "additionalProperties": true
          },
          "dryRunCommerce": {
            "type": "object",
            "additionalProperties": true
          },
          "publicDemo": {
            "type": "object",
            "additionalProperties": true
          },
          "protectedFlow": {
            "type": "object",
            "additionalProperties": true
          },
          "tools": {
            "type": "object",
            "additionalProperties": true
          },
          "truthLabels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TruthLabel"
            }
          },
          "truth": {
            "type": "string"
          }
        }
      },
...
      "AgentCommerceQuote": {
        "type": "object",
        "required": [
          "quoteId",
          "mode",
          "product",
          "siteWorldId",
          "sku",
          "quantity",
          "unitAmountCents",
          "totalAmountCents",
          "currency"
        ],
        "properties": {
          "quoteId": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "dry_run"
            ]
          },
          "product": {
            "$ref": "#/components/schemas/AgentCommerceProduct"
          },
          "siteWorldId": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "quantityLabel": {
            "type": "string"
          },
          "unitAmountCents": {
            "type": "integer"
          },
          "totalAmountCents": {
            "type": "integer"
          },
          "currency": {
            "type": "string",
            "enum": [
              "usd"
            ]
          },
          "entitlementType": {
            "type": "string",
            "enum": [
              "package_access",
              "hosted_session"
            ]
          },
          "truthLabels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TruthLabel"
            }
          }
        }
      },
      "AgentCommerceQuoteResponse": {
        "type": "object",
        "required": [
          "quote",
          "truth"
        ],
        "properties": {
          "quote": {
            "$ref": "#/components/schemas/AgentCommerceQuote"
          },
          "siteWorld": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          },
          "truth": {
            "type": "string"
          }
        }
      },
      "AgentDryRunCheckoutRequest": {
        "type": "object",
        "required": [
          "siteWorldId",
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "dry_run"
            ]
          },
          "siteWorldId": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/AgentCommerceProduct"
          },
          "sessionHours": {
            "type": "integer",
            "minimum": 1,
            "default": 1
          },
          "buyer": {
            "type": "object",
            "properties": {
              "uid": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          }
        }
      },
      "AgentDryRunOrder": {
        "type": "object",
        "required": [
          "id",
          "dry_run",
          "status",
          "payment_status",
          "fulfillment_status",
          "item",
          "pricing",
          "entitlement_id"
        ],
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "string"
          },
          "dry_run": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "fulfilled"
            ]
          },
          "payment_status": {
            "type": "string",
            "enum": [
              "dry_run_paid"
            ]
          },
          "fulfillment_status": {
            "type": "string",
            "enum": [
              "provisioned"
            ]
          },
          "item": {
            "type": "object",
            "additionalProperties": true
          },
          "pricing": {
            "type": "object",
            "additionalProperties": true
          },
          "stripe": {
            "type": "object",
            "additionalProperties": true
          },
          "entitlement_id": {
            "type": "string"
          }
        }
      },
      "AgentMarketplaceEntitlement": {
        "type": "object",
        "required": [
          "id",
          "order_id",
          "sku",
          "access_state"
        ],
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "string"
          },
          "order_id": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "item_type": {
            "$ref": "#/components/schemas/AgentCommerceProduct"
          },
          "delivery_mode": {
            "type": "string"
          },
          "access_state": {
            "type": "string",
            "enum": [
              "provisioned",
              "manual_review_required",
              "revoked"
            ]
          },
          "dry_run": {
            "type": "boolean"
          }
        }
      },
      "AgentDryRunOrderResponse": {
        "type": "object",
        "properties": {
          "quote": {
            "$ref": "#/components/schemas/AgentCommerceQuote"
          },
          "order": {
            "$ref": "#/components/schemas/AgentDryRunOrder"
          },
          "receipt": {
            "type": "object",
            "additionalProperties": true
          },
          "entitlement": {
            "$ref": "#/components/schemas/AgentMarketplaceEntitlement"
          },
          "truth": {
            "type": "string"
          }
        }
      },
      "AgentEntitlementResponse": {
        "type": "object",
        "required": [
          "entitlement"
        ],
        "properties": {
          "entitlement": {
            "$ref": "#/components/schemas/AgentMarketplaceEntitlement"
          }
        }
      },
      "AgentEntitlementReadiness": {
        "type": "object",
        "required": [
          "mode",
          "siteWorldId",
          "entitled",
          "launchable",
          "blockers"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "dry_run"
            ]
          },
          "siteWorldId": {
            "type": "string"
          },
          "entitlement": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          },
          "entitled": {
            "type": "boolean"
          },
          "launchable": {
            "type": "boolean"
          },
          "blockers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "truth": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "diagnostic": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  },
...
}